Rename AttrsField to ForwardedField
This provides a reusable primitive for all forwarded fields which should support `with`
