+++ core/src/options/mod.rs
@@ -7,6 +7,7 @@ use crate::{Error, FromMeta, Result};
 
 mod core;
 mod forward_attrs;
+mod forwarded_field;
 mod from_attributes;
 mod from_derive;
 mod from_field;
@@ -19,7 +20,8 @@ mod outer_from;
 mod shape;
 
 pub use self::core::Core;
-pub use self::forward_attrs::{AttrsField, ForwardAttrsFilter};
+pub use self::forward_attrs::ForwardAttrsFilter;
+pub use self::forwarded_field::ForwardedField;
 pub use self::from_attributes::FromAttributesOptions;
 pub use self::from_derive::FdiOptions;
 pub use self::from_field::FromFieldOptions;
